redo a bunch of things to fix all and forever borrowing shenanigans for python and others
