+++ src/error.rs
@@ -6,6 +6,7 @@ pub enum DataProcessingError {
     IncompatibleInplace(String),
     IncompatibleInputArray(String),
     InternalError(String),
+    MissingContext(String),
 }
 
 impl fmt::Display for DataProcessingError {
@@ -15,6 +16,7 @@ impl fmt::Display for DataProcessingError {
             DataProcessingError::IncompatibleInplace(e) => write!(f, "Unable to perform inplace operation! {}", e),
             DataProcessingError::IncompatibleInputArray(e) => write!(f, "Incompatible input array for images! {}", e),
             DataProcessingError::InternalError(e) => write!(f, "Internal error! Please report this! {}", e),
+            DataProcessingError::MissingContext(e) => write!(f, "Missing context! {}", e),
         }
     }
 }
